feat: display all products

+++ app/Domain/Product/Database/Migrations/2022_09_19_110245_create_categories_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the Migrations.
      *
@@ -17,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->unique();
             $table->string('slug')->unique()->index();
-            $table->decimal('discount', 3, 1)->nullable();
+            $table->integer('discount')->nullable();
             $table->timestamps();
         });
     }
